+++ tests/test_public_api.py
@@ -2,6 +2,7 @@
 
 import pytest
 
+import sunnbear._core.builtin_artifacts
 import sunnbear._core.data.exceptions
 import sunnbear._core.exceptions
 import sunnbear._core.functions.core
@@ -10,6 +11,7 @@ import sunnbear._core.solvers.bracketing
 import sunnbear._core.solvers.core
 import sunnbear._core.solvers.core.exceptions
 import sunnbear._core.stats
+import sunnbear.data
 import sunnbear.exceptions
 import sunnbear.functions
 import sunnbear.solvers
@@ -34,6 +36,7 @@ def _is_submodule(module, name: str) -> bool:
     "public_module, implementation_modules",
     [
         (sunnbear.stats, (sunnbear._core.stats,)),
+        (sunnbear.data, (sunnbear._core.builtin_artifacts,)),
         (sunnbear.functions, (sunnbear._core.functions.core,)),
         (sunnbear.solvers, (sunnbear._core.solvers.core, sunnbear._core.solvers.bracketing)),
         (